Номера строк больше не переполняются после int.MaxValue. Начало и шаг хранятся в настройках как int, но прежний внутренний счётчик тоже был int: при большом N и следующем кадре unchecked-сложение превращало положительный номер в отрицательный, и управляющая программа продолжалась строкой вида N-2. Счётчик форматтера и номер структурного блока переведены на long; при действующем пределе размера траектории его диапазона достаточно с большим запасом. Добавлена проверка границы: после N2147483647 следует N4294967294, то же число остаётся в модели блока. Release-тесты ядра: 547/547

+++ GCodeGenerator.Core/GCodeGenerators/GCodeFormatter.cs
@@ -31,7 +31,7 @@ namespace GCodeGenerator.GCodeGenerators
             // Пункт 8.1 плана: форматирование читает только группу Format.
             var format = settings.Format;
             var lines = new List<string>(program.Blocks.Count);
-            int lineNumber = format.UseLineNumbers ? format.LineNumberStart : 0;
+            long lineNumber = format.UseLineNumbers ? format.LineNumberStart : 0;
 
             foreach (var block in program.Blocks)
             {
